feature(home): Separate page and add bottom sheet
