+++ package/home/lib/home.dart
@@ -1,1 +1,2 @@
-
+export 'package:home/presentation/pages/home_page.dart';
+export 'package:home/presentation/pages/edit_profile_page.dart';
