+++ package/home/lib/presentation/pages/home_page.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:theme/theme.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'edit_profile_page.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({super.key});
@@ -37,13 +38,13 @@ class _HomePageState extends State<HomePage> {
                         icon: SvgPicture.asset(
                           'assets/icon/settings_24px.svg',
                           package: 'home',
+                          color: context.colors.onSurfaceVariant,
                         ),
                         onPressed: () {
                           Navigator.push(
                               context,
                               MaterialPageRoute(
-                                  builder: (context) =>
-                                      const EditProfileRoute()));
+                                  builder: (context) => EditProfilePage()));
                         })),
               ),
               SizedBox(height: 20),
@@ -177,7 +178,7 @@ class _HomePageState extends State<HomePage> {
                       child: InkWell(
                         splashColor: context.colors.surfaceVariant,
                         onTap: () {
-                          debugPrint('Daftar tugas Tapped.');
+                          debugPrint('Kalender Tapped.');
                         },
                         child: SizedBox(
                           height: 150,
@@ -223,7 +224,7 @@ class _HomePageState extends State<HomePage> {
                       child: InkWell(
                         splashColor: context.colors.surfaceVariant,
                         onTap: () {
-                          debugPrint('Daftar tugas Tapped.');
+                          debugPrint('Kegiatan Tapped.');
                         },
                         child: SizedBox(
                           height: 150,
@@ -266,7 +267,7 @@ class _HomePageState extends State<HomePage> {
                       child: InkWell(
                         splashColor: context.colors.surfaceVariant,
                         onTap: () {
-                          debugPrint('Daftar tugas Tapped.');
+                          debugPrint('Keluarga Tapped.');
                         },
                         child: SizedBox(
                           height: 150,
@@ -304,257 +305,6 @@ class _HomePageState extends State<HomePage> {
             ],
           ),
         ),
-      ),
-    );
-  }
-}
-
-enum MoodUser { SangatSenang, Senang, BiasaSaja, TidakSenang, Sakit }
-
-class EditProfileRoute extends StatefulWidget {
-  const EditProfileRoute({super.key});
-
-  @override
-  State<EditProfileRoute> createState() => _EditProfileRouteState();
-}
-
-class _EditProfileRouteState extends State<EditProfileRoute> {
-  MoodUser? _moodUser = MoodUser.SangatSenang;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Edit profil'),
-        actions: [IconButton(onPressed: () {}, icon: Icon(Icons.check))],
-      ),
-      body: Container(
-        margin: const EdgeInsets.symmetric(horizontal: 16),
-        child: Column(children: [
-          SizedBox(
-            height: 20,
-          ),
-          SvgPicture.asset(
-            'assets/image/Group 6.svg',
-            package: 'home',
-          ),
-          SizedBox(
-            height: 20,
-          ),
-          TextFormField(
-            textInputAction: TextInputAction.next,
-            decoration: InputDecoration(
-              border: const OutlineInputBorder(),
-              label: const Text('Nama'),
-              suffixIcon: IconButton(
-                icon: Icon(
-                  Icons.cancel_outlined,
-                  color: context.colors.onSurfaceVariant,
-                ),
-                tooltip: 'Delete',
-                onPressed: () {},
-              ),
-            ),
-          ),
-          SizedBox(
-            height: 20,
-          ),
-          TextFormField(
-            textInputAction: TextInputAction.next,
-            decoration: InputDecoration(
-              border: const OutlineInputBorder(),
-              label: const Text('Tanggal ulang tahun'),
-              suffixIcon: IconButton(
-                icon: Icon(
-                  Icons.today,
-                  color: context.colors.onSurfaceVariant,
-                ),
-                tooltip: 'Delete',
-                onPressed: () {},
-              ),
-            ),
-          ),
-          SizedBox(
-            height: 20,
-          ),
-          TextFormField(
-            textInputAction: TextInputAction.next,
-            decoration: InputDecoration(
-              border: const OutlineInputBorder(),
-              label: const Text('Status'),
-              suffixIcon: IconButton(
-                icon: Icon(
-                  Icons.cancel_outlined,
-                  color: context.colors.onSurfaceVariant,
-                ),
-                tooltip: 'Delete',
-                onPressed: () {},
-              ),
-            ),
-          ),
-          SizedBox(
-            height: 20,
-          ),
-          TextFormField(
-            textInputAction: TextInputAction.next,
-            decoration: InputDecoration(
-              border: const OutlineInputBorder(),
-              label: const Text('Mood'),
-              suffixIcon: IconButton(
-                icon: SvgPicture.asset(
-                  'assets/icon/add_reaction.svg',
-                  package: 'home',
-                  color: context.colors.onSurfaceVariant,
-                ),
-                tooltip: 'Delete',
-                onPressed: () => showDialog(
-                    context: context,
-                    builder: (BuildContext context) {
-                      return StatefulBuilder(builder: (context, setState) {
-                        return AlertDialog(
-                          title: Text("Pilih mood Anda",
-                              style: context.titleLarge?.copyWith(
-                                  color: context.colors.onSurfaceVariant)),
-                          content: Column(
-                            mainAxisSize: MainAxisSize.min,
-                            children: [
-                              ListTile(
-                                title: Row(children: [
-                                  SvgPicture.asset(
-                                    'assets/icon/sentiment_very_satisfied.svg',
-                                    package: 'home',
-                                  ),
-                                  Text(
-                                    ' Sangat Senang',
-                                    style: context.bodySmall?.copyWith(
-                                        color: context.colors.onSurfaceVariant),
-                                  )
-                                ]),
-                                leading: Radio(
-                                  value: MoodUser.SangatSenang,
-                                  groupValue: _moodUser,
-                                  onChanged: (MoodUser? value) {
-                                    setState(() {
-                                      _moodUser = value;
-                                    });
-                                  },
-                                ),
-                              ),
-                              ListTile(
-                                title: Row(children: [
-                                  SvgPicture.asset(
-                                    'assets/icon/sentiment_satisfied.svg',
-                                    package: 'home',
-                                  ),
-                                  Text(
-                                    ' Senang',
-                                    style: context.bodySmall?.copyWith(
-                                        color: context.colors.onSurfaceVariant),
-                                  )
-                                ]),
-                                leading: Container(
-                                  child: Radio(
-                                    value: MoodUser.BiasaSaja,
-                                    groupValue: _moodUser,
-                                    onChanged: (MoodUser? value) {
-                                      setState(() {
-                                        _moodUser = value;
-                                      });
-                                    },
-                                  ),
-                                ),
-                              ),
-                              ListTile(
-                                title: Row(children: [
-                                  SvgPicture.asset(
-                                    'assets/icon/sentiment_neutral.svg',
-                                    package: 'home',
-                                  ),
-                                  Text(
-                                    ' Biasa aja',
-                                    style: context.bodySmall?.copyWith(
-                                        color: context.colors.onSurfaceVariant),
-                                  )
-                                ]),
-                                leading: Radio(
-                                  value: MoodUser.TidakSenang,
-                                  groupValue: _moodUser,
-                                  onChanged: (MoodUser? value) {
-                                    setState(() {
-                                      _moodUser = value;
-                                    });
-                                  },
-                                ),
-                              ),
-                              ListTile(
-                                title: Row(children: [
-                                  SvgPicture.asset(
-                                    'assets/icon/sentiment_dissatisfied.svg',
-                                    package: 'home',
-                                  ),
-                                  Text(
-                                    ' Tidak senang',
-                                    style: context.bodySmall?.copyWith(
-                                        color: context.colors.onSurfaceVariant),
-                                  )
-                                ]),
-                                leading: Radio(
-                                  value: MoodUser.Senang,
-                                  groupValue: _moodUser,
-                                  onChanged: (MoodUser? value) {
-                                    setState(() {
-                                      _moodUser = value;
-                                    });
-                                  },
-                                ),
-                              ),
-                              ListTile(
-                                title: Row(children: [
-                                  SvgPicture.asset(
-                                    'assets/icon/sick.svg',
-                                    package: 'home',
-                                  ),
-                                  Text(
-                                    ' Sakit',
-                                    style: context.bodySmall?.copyWith(
-                                        color: context.colors.onSurfaceVariant),
-                                  )
-                                ]),
-                                leading: Radio(
-                                  value: MoodUser.Sakit,
-                                  groupValue: _moodUser,
-                                  onChanged: (MoodUser? value) {
-                                    setState(() {
-                                      _moodUser = value;
-                                    });
-                                  },
-                                ),
-                              ),
-                            ],
-                          ),
-                          actions: [
-                            TextButton(
-                              onPressed: () => Navigator.pop(context, 'Cancel'),
-                              child: Text('Cancel',
-                                  style: context.bodyMedium?.copyWith(
-                                      color: context.colors.primary)),
-                            ),
-                            TextButton(
-                              onPressed: () => Navigator.pop(context, 'OK'),
-                              child: Text(
-                                'OK',
-                                style: context.bodyMedium
-                                    ?.copyWith(color: context.colors.primary),
-                              ),
-                            ),
-                          ],
-                        );
-                      });
-                    }),
-              ),
-            ),
-          ),
-        ]),
       ),
     );
   }
