+++ pubspec.yaml
@@ -36,13 +36,13 @@ dependencies:
     home:
         path: package/home
 
-
     # The following adds the Cupertino Icons font to your application.
     # Use with the CupertinoIcons class for iOS style icons.
     cupertino_icons: ^1.0.2
     dynamic_color: ^1.5.4
     firebase_core: ^2.3.0
     firebase_crashlytics: ^3.0.6
+    go_router: ^5.2.1
 
 dev_dependencies:
     flutter_test:
